Change error logic and handling
Change the error logic to test for the most likely outcome, instead of
least likely. Because QSslError::SslError is an enum declared in a
class, we need to change the error handling to prevent to many
creations of this class and its associated overhead. So we pass a
pointer to QSslError to the function we are calling and check QSslError
to see if an occured after calling the function.

Change-Id: Id50a31dd51ef7d6e6a0f931707769182932c100a
Signed-off-by: Martin Banky <Martin.Banky@gmail.com>